"Agario" uses PFGA and is written in verilog
This project is inspired by the popular computer game Agario. In this
version of the game, the player is represented by a square, and needs
to dodge larger squares, while eating smaller squares. When the player
collides with a square of a bigger size, the player gets eaten and the
game ends. On the other hand, if the size of the player is bigger than
the size of the square in the collision, the player grows bigger and
gains score. The player also gains passive points for staying alive. In
gameplay, the user uses the arrow keys on the keyboard to move, and can
toggle the switches of the DE1-SoC board to change the color of the
enemies.